+++ frontend/src/index.js
@@ -26,8 +26,6 @@ const App = () => {
     <Routes>
       <Route path="/" element={<Pages.Landing />} />
       <Route path="/editor" element={<Pages.Editor />} />
-      <Route path="/login" element={<Pages.Login />} />
-      <Route path="/signup" element={<Pages.Signup />} />
       <Route path="/logout" element={<Pages.Logout />} />
       <Route path="/about" element={<Pages.About />} />
       <Route path="/privacy" element={<Pages.Privacy />} />
