+++ frontend/src/pages/Login/Login.js
@@ -1,8 +1,7 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useRef, forwardRef } from 'react'
 import { useForm } from 'react-hook-form'
-import { useNavigate, useLocation } from 'react-router-dom'
 
-import { Main, TextInput, Button, Label, Header } from '/src/components'
+import { Main, TextInput, Button, Label, Header, Modal } from '/src/components'
 import { useAuth } from '/src/hooks'
 
 const defaultValues = {
@@ -10,12 +9,12 @@ const defaultValues = {
   password: '',
 }
 
-const Login = () => {
-  const location = useLocation()
-  const navigate = useNavigate()
+const Login = {}
+
+Login.Form = forwardRef(({ setFormActions, onComplete, ...props }, ref) => {
   const [fireError, setFireError] = useState(null)
   const [isSubmitting, setIsSubmitting] = useState(false)
-  const { user: currentUser, isLoading, signIn, error: authError } = useAuth()
+  const { isLoading, signIn, error: authError } = useAuth()
 
   const {
     register,
@@ -23,18 +22,16 @@ const Login = () => {
     formState: { errors, isDirty },
   } = useForm({ defaultValues})
 
-  // Determine previous page (if there was such a thing)
-  const previousPage = location?.state?.from
-
   // Were there errors from fire / from signing in?
   const error = fireError || authError?.message
 
-  // Navigate away if already logged in
   useEffect(() => {
-    if (currentUser && !isSubmitting) {
-      navigate('/new')
+    if (setFormActions) {
+      setFormActions(<>
+        <Button type='submit' form='login-form' disabled={!isDirty || isLoading || isSubmitting}>Login</Button> 
+      </>)
     }
-  }, [currentUser])
+  }, [isDirty, isLoading, isSubmitting, ref?.current, setFormActions])
 
   const onSubmit = async values => {
     setIsSubmitting(true)
@@ -42,34 +39,46 @@ const Login = () => {
     signIn(values.email, values.password)
       .then(() => {
         setIsSubmitting(false)
-        if (previousPage)
-          navigate(previousPage)
       })
-      .then(() => navigate('/new'))
+      .then(() => { onComplete?.() })
       .catch(() => {
         setFireError('Incorrect email or password')
         setIsSubmitting(false)
       })
   }
 
-  return <Main>
+  return <form onSubmit={handleSubmit(onSubmit)} ref={ref} id='login-form' {...props}>
+    {error && (
+      <p>{error}</p>
+    )}
+    <Label htmlFor='login-email'>Email</Label>
+    <TextInput id='login-email' type='email' {...register('email')} />
+    <p>{errors.email?.message}</p>
+
+    <Label htmlFor='login-password'>Password</Label>
+    <TextInput id='login-password' type='password' {...register('password')} />
+    <p>{errors.email?.password}</p>
+  </form>
+})
+
+Login.Modal = ({ ...props }) => {
+  const [formActions, setFormActions] = useState()
+  
+  return <Modal
+    narrow
+    actions={<>
+      <Button secondary style={{ marginRight: 'auto' }} onClick={props?.onClose}>Close</Button> 
+      {formActions}
+    </>}
+    {...props}
+  >
     <Header center/>
     <h2>Login</h2>
-    <form onSubmit={handleSubmit(onSubmit)}>
-      {error && (
-        <p>{error}</p>
-      )}
-      <Label htmlFor='email'>Email</Label>
-      <TextInput type='email' {...register('email')} />
-      <p>{errors.email?.message}</p>
-
-      <Label htmlFor='password'>Password</Label>
-      <TextInput type='password' {...register('password')} />
-      <p>{errors.email?.password}</p>
-
-      <Button type='submit' disabled={!isDirty || isSubmitting}>Login</Button>
-    </form>
-  </Main>
+    <Login.Form
+      onComplete={props?.onClose}
+      setFormActions={setFormActions}
+      style={{ paddingBottom: '1em' }} />
+  </Modal>
 }
 
 export default Login
